Pollard's Rho algorithm for breaking DLP (#28)
Introduce `Pollard's Rho` algorithm for breaking DLP.

Implemented Pollard's rho algorithm and its test, taking secp256k1 as an
example. However, Pollard's rho algorithm takes a _very long_ time;
therefore, I added an `ignore` tag to tests.


issue: #18

+++ elliptic-curves-rs/src/breaking_dlp/brute_force.rs
@@ -17,7 +17,7 @@ impl<C: Curve> DiscreteLog<C> for BruteForce {
     fn solve(
             p: &CurvePoint<C>,
             q: &CurvePoint<C>,
-        ) -> (u16, C::ScalarField) {
+        ) -> (u64, C::ScalarField) {
         if !C::is_on_curve(&p.inner) || !C::is_on_curve(&q.inner) {
             panic!("p or q are not on curve");
         }
@@ -25,7 +25,7 @@ impl<C: Curve> DiscreteLog<C> for BruteForce {
         let mut rng = thread_rng();
         let start = C::ScalarField::rand(&mut rng);
         let mut current = start;
-        let end = u16::MAX;
+        let end = u64::MAX;
 
         for steps in 0..end {
             let result = p.mul_scalar(&current);
